requirements.txt
seaborn==0.11.1
pygad==2.16.0
pandas==1.3.0
tqdm==4.61.2
Biopython==1.79
numpy==1.21.0
matplotlib==3.4.2